refactor: split shared models
