+++ TheBackend.Admin/_Imports.razor
@@ -9,4 +9,4 @@
 @using TheBackend.Admin
 @using TheBackend.Admin.Layout
 @using TheBackend.Admin.Shared
-@using TheBackend.DynamicModels
+@using TheBackend.SharedModels
