+++ TheBackend.Api/Controllers/ModelsController.cs
@@ -1,4 +1,5 @@
-﻿using TheBackend.DynamicModels;
+using TheBackend.DynamicModels;
+using TheBackend.SharedModels;
 using Microsoft.AspNetCore.Mvc;
 using TheBackend.Api;
