+++ TheBackend.SharedModels/Models/Customer.cs
@@ -3,14 +3,15 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-namespace TheBackend.DynamicModels
+
+namespace TheBackend.SharedModels
 {
     public class Customer
     {
         public int CustomerId { get; set; }
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
-        public string Email { get; set; }
+        public string FirstName { get; set; } = string.Empty;
+        public string LastName { get; set; } = string.Empty;
+        public string Email { get; set; } = string.Empty;
         public bool IsActive { get; set; }
     }
 }
