+++ TheBackend.SharedModels/Models/Product.cs
@@ -3,12 +3,13 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-namespace TheBackend.DynamicModels
+
+namespace TheBackend.SharedModels
 {
-    public class User
+    public class Product
     {
         public int Id { get; set; }
-        public string Name { get; set; }
-        public string Email { get; set; }
+        public string Name { get; set; } = string.Empty;
+        public decimal Price { get; set; }
     }
 }
