+++ TheBackend.sln
@@ -17,6 +17,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TheBackend.DynamicModels", 
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TheBackend.Admin", "TheBackend.Admin\TheBackend.Admin.csproj", "{4F8B9237-517F-42D9-9FFC-7469A1188A40}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TheBackend.SharedModels", "TheBackend.SharedModels\TheBackend.SharedModels.csproj", "{1F4B65DC-64EE-4733-BA78-30121D4E50D3}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -111,6 +113,18 @@ Global
 		{4F8B9237-517F-42D9-9FFC-7469A1188A40}.Release|x64.Build.0 = Release|Any CPU
 		{4F8B9237-517F-42D9-9FFC-7469A1188A40}.Release|x86.ActiveCfg = Release|Any CPU
 		{4F8B9237-517F-42D9-9FFC-7469A1188A40}.Release|x86.Build.0 = Release|Any CPU
+		{1F4B65DC-64EE-4733-BA78-30121D4E50D3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1F4B65DC-64EE-4733-BA78-30121D4E50D3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1F4B65DC-64EE-4733-BA78-30121D4E50D3}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{1F4B65DC-64EE-4733-BA78-30121D4E50D3}.Debug|x64.Build.0 = Debug|Any CPU
+		{1F4B65DC-64EE-4733-BA78-30121D4E50D3}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{1F4B65DC-64EE-4733-BA78-30121D4E50D3}.Debug|x86.Build.0 = Debug|Any CPU
+		{1F4B65DC-64EE-4733-BA78-30121D4E50D3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1F4B65DC-64EE-4733-BA78-30121D4E50D3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1F4B65DC-64EE-4733-BA78-30121D4E50D3}.Release|x64.ActiveCfg = Release|Any CPU
+		{1F4B65DC-64EE-4733-BA78-30121D4E50D3}.Release|x64.Build.0 = Release|Any CPU
+		{1F4B65DC-64EE-4733-BA78-30121D4E50D3}.Release|x86.ActiveCfg = Release|Any CPU
+		{1F4B65DC-64EE-4733-BA78-30121D4E50D3}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
